Make: move queue_screen.h to include directory
